Rename CityMotion Plugin and release 1.2.7 location updates
