+++ iOS-XCode/CMPlugin.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:..">
+      location = "self:../CMPlugin">
    </FileRef>
 </Workspace>
