+++ iOS-XCode/CMPlugin.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
