+++ iOS-XCode/CMPlugin/Assets.xcassets/Contents.json
