+++ iOS-XCode/CMPlugin/Info.plist
